Fixed bugs in algorithm for check parameters. Insert function definitions into hash table

+++ lex_analyzer.h
@@ -19,8 +19,8 @@
 
 typedef struct buffer_struct
 {
-    unsigned long long int max_length;
-    unsigned long long int position;
+    size_t max_length;
+    size_t position;
     char* data;
 }* BUFFER_STRUCT;
